Better implementation. Better command line parser. Diferent output types (vcd or csv)

+++ src/vcoscillator.cpp
@@ -1,5 +1,5 @@
 #define _USE_MATH_DEFINES
-#include <math.h>
+#include <cmath>
 #include "vcoscillator.h"
 
 using namespace std;
@@ -8,10 +8,7 @@ void vcoscillator::processing(void) {
 	double frequency = kvo*vc.read() + fmin;
 	double time = sc_time_stamp().to_seconds();
 
-	//freq.write(((vc.read()/vcc)-0.5)*BW+F0);
-	//out.write( VGND+VGND*sin( sc_time_stamp().to_seconds()*((((vc.read()/VCC)-0.5)*BW+F0)*2.*M_PI))) ;
-
-	out.write((double)(vcm+vcm*sin(2*M_PI*frequency*time)));
+	out.write((double)(vcm+vcm*(double)sin(double(2*M_PI*frequency*time))));
 	freq.write(frequency);
 }
